Refactor components to split
